+++ admob-latest/src/test/java/com/rakuten/tech/mobile/miniapp/ads/TestData.kt
@@ -1,0 +1,3 @@
+package com.rakuten.tech.mobile.miniapp
+
+internal const val TEST_AD_UNIT_ID = "ca-app-pub-xxxxxxxxxxxxxxxx/xxxxxxxxxx"