+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/userdata/ContactListActivity.kt
@@ -19,23 +19,28 @@ import androidx.recyclerview.widget.RecyclerView
 import com.rakuten.tech.mobile.miniapp.js.userinfo.Contact
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.ContactsActivityBinding
-import com.rakuten.tech.mobile.testapp.helper.defaultContact
 import com.rakuten.tech.mobile.testapp.helper.isEmailValid
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
+import java.security.SecureRandom
+import java.util.Locale
 import java.util.UUID
+import kotlin.collections.ArrayList
 
-class ContactListActivity : BaseActivity() {
+class ContactListActivity : BaseActivity(), ContactListener {
 
     private lateinit var settings: AppSettings
     private lateinit var binding: ContactsActivityBinding
-    private val adapter = ContactListAdapter()
+    private val adapter = ContactListAdapter(this)
     private var contactListPrefs: SharedPreferences? = null
     private var isFirstLaunch: Boolean
         get() = contactListPrefs?.getBoolean(IS_FIRST_TIME, true) ?: true
         set(value) {
             contactListPrefs?.edit()?.putBoolean(IS_FIRST_TIME, value)?.apply()
         }
+
+    private val fakeFirstNames = arrayOf("Yvonne", "Jamie", "Leticia", "Priscilla", "Sidney", "Nancy", "Edmund", "Bill", "Megan")
+    private val fakeLastNames = arrayOf("Andrews", "Casey", "Gross", "Lane", "Thomas", "Patrick", "Strickland", "Nicolas", "Freeman")
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -45,7 +50,7 @@ class ContactListActivity : BaseActivity() {
         settings = AppSettings.instance
         showBackIcon()
         binding = DataBindingUtil.setContentView(this, R.layout.contacts_activity)
-        renderContactList()
+        renderRandomContactList()
         binding.fabAddContact.setOnClickListener { onAddAction() }
     }
 
@@ -75,46 +80,81 @@ class ContactListActivity : BaseActivity() {
     }
 
     private fun onAddAction() {
+        showDialog(isUpdate = false)
+    }
+
+    private fun showDialog(isUpdate: Boolean, position: Int? = null){
         val contactView = layoutInflater.inflate(R.layout.dialog_add_contact, null)
         val edtContactId = contactView.findViewById<AppCompatEditText>(R.id.edtContactId)
         val edtContactName = contactView.findViewById<AppCompatEditText>(R.id.edtContactName)
         val edtContactEmail = contactView.findViewById<AppCompatEditText>(R.id.edtContactEmail)
 
         ContactInputDialog.Builder().build(this).apply {
+            val randomContact = createRandomContact()
+            edtContactId.setText(randomContact.id)
+            edtContactName.setText(randomContact.name)
+            edtContactEmail.setText(randomContact.email)
+
             setView(contactView)
+
+            if (isUpdate) {
+                setPositiveButton(getString(R.string.action_update))
+                position?.let {
+                    val existingContact = adapter.provideContactEntries()[it]
+                    edtContactId.setText(existingContact.id)
+                    edtContactName.setText(existingContact.name)
+                    edtContactEmail.setText(existingContact.email)
+                    edtContactId.isEnabled = false
+                }
+            } else {
+                setPositiveButton(getString(R.string.action_add))
+                edtContactId.isEnabled = true
+            }
+
             setPositiveListener(View.OnClickListener {
-                var canSave = true
                 val id: String = edtContactId.text.toString().trim()
-                var name: String? = edtContactName.text.toString().trim()
-                var email: String? = edtContactEmail.text.toString().trim()
+                val name: String = edtContactName.text.toString().trim()
+                val email: String = edtContactEmail.text.toString().trim()
 
-                if (id.isNotEmpty()) {
-                    if (name!!.isEmpty())
-                        name = null
-
-                    if (email!!.isEmpty())
-                        email = null
-                    else {
-                        if (!email.isEmailValid()) {
-                            canSave = false
-                            Toast.makeText(this@ContactListActivity, getString(R.string.userdata_error_invalid_contact_email), Toast.LENGTH_LONG)
-                                .apply { setGravity(Gravity.TOP, 0, 50) }
-                                .show()
-                        }
-                    }
-                } else {
-                    canSave = false
-                    Toast.makeText(this@ContactListActivity, getString(R.string.userdata_error_invalid_contact_id), Toast.LENGTH_LONG)
-                        .apply { setGravity(Gravity.TOP, 0, 50) }
-                        .show()
+                if (isVerifiedContact(id, name, email)) {
+                    val contact = Contact(id = id, name = name, email = email)
+                    if (isUpdate) position?.let { adapter.updateContact(it, contact) }
+                    else adapter.addContact(adapter.itemCount, contact)
                 }
 
-                if (canSave) {
-                    adapter.addContact(adapter.itemCount, Contact(id = id, name = name, email = email))
-                    this.dialog?.cancel()
-                }
+                this.dialog?.cancel()
             })
         }.show()
+    }
+
+    private fun isVerifiedContact(id: String, name: String, email: String): Boolean {
+        var isVerified = true
+
+        if (id.isEmpty()) {
+            isVerified = false
+            showContactInputWarning(getString(R.string.userdata_error_empty_contact_id))
+        }
+
+        if (name.isEmpty()) {
+            isVerified = false
+            showContactInputWarning(getString(R.string.userdata_error_empty_contact_name))
+        }
+
+        if (email.isNotEmpty() && !email.isEmailValid()) {
+            isVerified = false
+            showContactInputWarning(getString(R.string.userdata_error_invalid_contact_email))
+        } else if (email.isEmpty()) {
+            isVerified = false
+            showContactInputWarning(getString(R.string.userdata_error_empty_email_name))
+        }
+
+        return isVerified
+    }
+
+    private fun showContactInputWarning(message: String) {
+        Toast.makeText(this@ContactListActivity, message, Toast.LENGTH_LONG)
+                .apply { setGravity(Gravity.TOP, 0, 100) }
+                .show()
     }
 
     override fun onResume() {
@@ -122,7 +162,7 @@ class ContactListActivity : BaseActivity() {
         if (isFirstLaunch) isFirstLaunch = false
     }
 
-    private fun renderContactList() {
+    private fun renderRandomContactList() {
         if (!isFirstLaunch && settings.isContactsSaved) {
             if (settings.contacts.isEmpty()) {
                 renderAdapter(arrayListOf())
@@ -134,8 +174,16 @@ class ContactListActivity : BaseActivity() {
     }
 
     private fun createRandomContactList(): ArrayList<Contact> = ArrayList<Contact>().apply {
-        for (i in 1..10)
-            this.add(defaultContact(UUID.randomUUID().toString().trimEnd()))
+        for (i in 1..10) {
+            this.add(createRandomContact())
+        }
+    }
+
+    private fun createRandomContact(): Contact {
+        val firstName = fakeFirstNames[(SecureRandom().nextDouble() * fakeFirstNames.size).toInt()]
+        val lastName = fakeLastNames[(SecureRandom().nextDouble() * fakeLastNames.size).toInt()]
+        val email = firstName.toLowerCase(Locale.ROOT) + "." + lastName.toLowerCase(Locale.ROOT) + "@example.com"
+        return Contact(UUID.randomUUID().toString().trimEnd(), "$firstName $lastName", email)
     }
 
     private fun renderAdapter(contactNames: ArrayList<Contact>) {
@@ -189,5 +237,9 @@ class ContactListActivity : BaseActivity() {
         fun start(activity: Activity) {
             activity.startActivity(Intent(activity, ContactListActivity::class.java))
         }
+    }
+
+    override fun onContactItemClick(position: Int) {
+        showDialog(isUpdate = true, position = position)
     }
 }